feat: increase savegame version to 2.0, greately reduce its size

+++ Assets/Modules/GameSession/Scripts/Model/BossInfo.cs
@@ -2,7 +2,7 @@
 {
 	public readonly partial struct BossInfo
 	{
-		public BossInfo(int count, long time)
+		public BossInfo(int count, int time)
 		{
 			_lastDefeatTime = time;
 			_defeatCount = count;
